Fix other declaration files
