+++ js/modules/signal.d.ts
@@ -1,0 +1,9 @@
+import { SecretSessionCipherConstructor } from './metadata/SecretSessionCipher';
+
+interface Metadata {
+  SecretSessionCipher: SecretSessionCipherConstructor;
+}
+
+export interface SignalInterface {
+  Metadata: Metadata;
+}
