--- ts/window/types/signal.ts
@@ -1,9 +1,0 @@
-import { SecretSessionCipher } from './SecretSessionCipher';
-
-interface Metadata {
-  SecretSessionCipher: typeof SecretSessionCipher;
-}
-
-export interface SignalInterface {
-  Metadata: Metadata;
-}
